Refactor schedules out of calendar and reset form after successful submit.
